+++ Module1/SeparatedCharLexer.cs
@@ -1,0 +1,95 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LexerTasks
+{
+    public class SeparatedCharsLexer : Lexer
+    {
+        protected System.Text.StringBuilder mesString;
+        public String message;
+
+        public SeparatedCharsLexer(string input)
+            : base(input)
+        {
+            mesString = new System.Text.StringBuilder();
+            message = "";
+        }
+
+        public override void Parse()
+        {
+            NextCh();
+
+            while (true)
+            {
+                if (char.IsLetter(currentCh))
+                {
+                    message += currentCh;
+                    NextCh();
+                }
+                else if (currentCh == ',' || currentCh == ';')
+                {
+                    NextCh();
+                    break;
+                }
+                else
+                {
+                    Error();
+                }
+            }
+
+            while (char.IsLetter(currentCh))
+            {
+                message += currentCh;
+                NextCh();
+            }
+
+            if (currentCharValue != -1)
+            {
+                Error();
+            }
+
+            System.Console.WriteLine("string with separated chars by ; or , is recognised " + message);
+        }
+
+        public static void Testing()
+        {
+            var tests = new Dictionary<string, string>{
+                { "a;", "a" },
+                { ";fr", "fr"},
+                { "abg,abg", "abgabg"},
+                { "abg;;", "error"},
+                { ",,", "error"},
+                { "tl;dr", "tldr"},
+                { ",glO", "glO"}
+            };
+
+            foreach (var t in tests)
+            {
+                var L = new SeparatedCharsLexer(t.Key);
+                bool passed = false;
+                try
+                {
+                    L.Parse();
+                    passed = L.message.Equals(t.Value);
+                }
+                catch (LexerException e)
+                {
+                    passed = true;
+                }
+
+                if (passed)
+                {
+                    System.Console.WriteLine("Test is passed");
+                }
+                else
+                {
+                    System.Console.WriteLine("Test is not passed");
+                }
+            }
+
+        }
+    }
+}
